Fix: Worked around Xcode availability nonsense
For subscript usage of `MDLVoxelAsset`, Xcode is insisting that `error: 'subscript' is only available on OS X 10.13 or newer`, which is true for `MDLVoxelAsset`'s ancestor type `MDLObjectContainerComponent` but not for our custom subclass `MDLVoxelAsset` _(since we're providing our own implementation, styled after `MDLAsset`'s implementation)_.

Tried to fix to no avail by creating Obj-C and/or Swift extensions to `MDLVoxelAsset` and/or `MDLObjectContainer` to provide the subscript method even when not using macOS 10.13, with and without a variety of availability macros.

Ended up just side-stepping the issue (for now) in the VoxViewer app by using the `.object(at: _)` function instead of directly subscripting.  Subscripting should still work for app projects using the `MDLVoxelAsset` lib with Swift code with a minimum deployment target of macOS 10.13 or newer, or Objective-C code of any stripe (Obj-C seems to be a bit saner about our overridden method).