Moved mesh generation into new MDLVoxelAssetModel class, one per .vox model
First-pass work-in-progress cleaving of the mesh-generation stuff in a new `MDLVoxelAssetModel` class, and `MDLVoxelAsset` just managing the public API and building a private array of `MDLVoxelAssetModel` to handle the mesh generation.

Sloppy, but it works.  Needs lots more massaging— the header inclusion is problematic, there's redundant code all over, and some of `MDLVoxelAsset`'s API is probably broken.  Some of the decisions on how to clean this up will become cleaner once we have per-model transforms— right now it's just putting all the `MDLMesh`s on the same `MDLObjectContainer` node.

Not going to go into all the details of what was done here because it's messy and amounted to:

1. _copy-and-pasted `MDLVoxelAsset` to new `MDLVoxelAssetModel.h`/`.m` files_
2. _deleted stuff that wasn't needed in each_
3. _added `_models` array ivar to `MDLVoxelAsset` and hooked it up_
4. _fixed a bunch of compiler errors due to duplicate and missing symbols_