Renamed consts to be more sane, consistent, & readable
Since I was getting confused as to what data certain constants are for, switched to a new naming scheme with 3 parts separated by underscores— 1. class or other grouping name, 2. variable or data piece, and 3. type of data on the constant is providing (e.g. size, offset, etc.).  Also made the 2nd part match the casing of the var it's pertaining to (so generally lower camel case), and the 3rd lowercase in most situations (except when it's stating a typename).

The end result is going from confusing names like `kVoxelChunkVoxels_Offset` to `kVoxelChunk_voxels_offset` (offset of `voxels` in `VoxelChunk`-related code).