Fix: Workaround for SceneKit console spam “Error: Pass FloorPass is not linked to …”
Subverts SceneKit bug that causes “Error: Pass FloorPass is not linked to the rendering graph and will be ignored check it's input/output” console spam (and reduced framerates) when `reflectivity` is set to `0.0`.

Instead, let it render barely visible, at very low resolution, with a bitmask that include nothing.  Still will cause a perf hit, but that's just what you get with SceneKit. ¯\_(ツ)_/¯